xz1919: merging IP-16 with develop
